Node Space: Small feature for Venomgfx, Shift + F select node of the same type
This is a small request from Venomgfx, select a node
and then press Shift + F to select all the nodes of the
same type (of the active node).

The key binding can be change, we thing in a "Find Next" (that is
way the FKEY) with venomgfx, but no problem with change that.

Also I add the entry in the select menu.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@28736 954f8c5b-7b00-dc11-b283-0030488c597c