Fix bug #22153 and #21609
Can't leave edit mode if you enter to edit mode and hide the
object from the outliner.

Also fix the problem if you hide the object and enter edit mode
from the outliner.

To avoid this problem you can't enter edit mode from the outliner
if the object is not visible and also you can't hide the object
from the outliner if the object is in edit mode.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@29222 954f8c5b-7b00-dc11-b283-0030488c597c