patch from Mitchell Stokes with some changes. BGE Py Controllers were effectively doing this...  "a.b.c" --> "__import__('a').b.c()"
This was annoying because it meant module 'a' would need to import 'b' explicitly.

Now use import like this.
 "a.b.c" --> "__import__("a.b").c()"

Note that this has the slight disadvantage that these need to be modules, where as before they could be collections of functions in a class instance for eg. So its possible this breaks existing files but dont think anyone used this since its a fairly obscure use case.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@30746 954f8c5b-7b00-dc11-b283-0030488c597c