Apply patch [#23745] fix #23078 operator panel disappears. Patch by Shane Ambler Reported by Anthony Edlin
From the patch details:
"
In response to bug #23078 the operator panel disappears when dragged above the top of the 3dview without showing the
azone icon to restore it.

This patch properly hides the operator panel if dragged near the top of the 3dview so that the azone icon is in place
"

Thanks!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@31986 954f8c5b-7b00-dc11-b283-0030488c597c