Fix [#21084] Collada messing up geometry on import and renaming uvlayers Reported by Michael Williamson
Add support to write the uvlayer name as found in <bind_vertex_input> semantic attribute (Blender
exports in this one too, so reimport goes ok now).

I have created a patch for OpenCOLLADA to be applied against r769, which is needed
for this fix to work: http://code.google.com/p/opencollada/issues/detail?id=117

The Windows OpenCOLLADA updated builds (r769+patches) will be committed after this.

Note: This means that linux and osx users will have to patch OpenCOLLADA themselves, until
it's committed (or in case of OSX the libs updated in lib/darwinXXX).

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@32235 954f8c5b-7b00-dc11-b283-0030488c597c