Fixed bug #23826, Other kind of brushes appear in sculpt mode
Was another problem caused by each brush being allowed in more than
one paint mode.

Added a new field to the brush struct to indicate what mode the icon
was last set for; if it's changed then reset it. Not sure if it's
really worth it to cache this, could remove it for simplicity.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@33647 954f8c5b-7b00-dc11-b283-0030488c597c