Todo item:
Object color option now is incorporated in preview render, and
in display in 3D window.

Note that Object-color is "modulating" the render result, it just
multiplies... which is very limited for practical use. This was
added for Apricot game project. The original meaning however was
to replace diffuse only... so specular and light work still correct.

ALso made header for properties editor start on top, to reflect
default (report in tracker).



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@34101 954f8c5b-7b00-dc11-b283-0030488c597c