"Available" Keying Set bugfix:
This builtin Keying Set is supposed to insert keyframes for every
F-Curve that exists for the selected data (usually objects and/or
bones only). However, as coded, it was only useful for objects, since
it would just go through all the F-Curves for the object's action,
instead of just the F-Curves relevant to a selected bone.

Tweaked the code to make this case (and similar ones) hopefully work
better by default.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@34222 954f8c5b-7b00-dc11-b283-0030488c597c