Added operator to show all armature layers (similar to the 3D View "Show All Layers"). This has been mapped to Ctrl-Accentkey
If necessary, you can alter your keymaps so that this operator is
invoked with its "all" property disabled. This will only toggle the
first row (first 16) layers, which is useful in most rigs for only
enabling all the layers with rig controls and not showing the layers
with rig mechanics.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@34264 954f8c5b-7b00-dc11-b283-0030488c597c