Bugfix [#25907] Impossible to "select all" on narrow action editors
Marker selection operators now use a special poll() callback which
checks that there are some markers before trying to run them. This
means that when there are no markers, the full screen range is
available, instead of just those areas above the region masked off for
markers which can be quite slim when trying to save space by only
showing the summary channel.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@34618 954f8c5b-7b00-dc11-b283-0030488c597c