Bugfix [#25934] "Apply Visual Transform as Pose" not working?
Simplified the code for Apply Visual Transform as Pose.

This makes it work for IK chains now, since there aren't any
recalculations of the pose involved now, which were contributing to
the pose getting incorrect as the applying process went on, since the
IK chain was getting calculated differently with each successive
change to the chain.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@34683 954f8c5b-7b00-dc11-b283-0030488c597c