Bugfix [#25987] Duplicated markers naming issue
One-liner fix - a missing "OPERATOR_FINISHED" on the select operator
was causing problems renaming markers and potentially with other
operations too!

To find this bug, I added debug method to dump the list of markers to
console. This has revealed some troublesome things about the way
markers are organised, which IMO need to be addressed.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@34727 954f8c5b-7b00-dc11-b283-0030488c597c