[#26476] <specular> and <shininess> missing from Colada reported by Juan Linietsky
Export <specular> for <phong> and <blinn> shaders, <shininess> was already being written for these.
<lambert> shader doesn't have <shininess>.

Right now we write <phong> when blender spec is phong, <blinn> when blender spec is blinn. When spec is
any other shader, and diffuse shader set to lambert, we export as <lambert>. Any other combination defaults
right now to <phong>. This will change when Blender specific profiles have been created for the shader
combinations in Blender.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@35705 954f8c5b-7b00-dc11-b283-0030488c597c