bugfix [#27091] Add new vertex at wrong position ( bpy.ops.mesh.dupli_extrude_cursor() ) 2
Ctrl+Click on mesh or curve view was using the selected points location or the cursors.
if either of these was behind the view it would add the point at (0, 0, 0).

now fallback to the view orbit pivot, added this option as an argument to view3d_get_view_aligned_coordinate().


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@36279 954f8c5b-7b00-dc11-b283-0030488c597c