Patch [#27344] 300+ New/changed tooltips by Davis Sorenson (dsavi). Thanks a lot! Reviewed by Tom Musgrove and myself.
From the patch description:
ValterVB on #blendercoders submitted a long list of missing tooltips in Blender, and I went through the list and added all I knew. After that I crowdsourced the rest by putting a spreadsheet on Google docs and having people fill in the missing ones that I didn't know. So if there's some weird tooltip in there that doesn't make sense, that's why.

Thanks to Wolter, spacetug and others on BlenderArtists for contributing tooltips.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@36704 954f8c5b-7b00-dc11-b283-0030488c597c