Backward compatibility fix for SOCK_DYNAMICS flag on group sockets.
This is currently only needed for displaying the up/down buttons of group sockets. All regular group sockets should have this flag to indicate they are added by the user. More complex "group-type" trees may use non-dynamic sockets in the future for sockets that are not supposed to be manipulated.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@39965 954f8c5b-7b00-dc11-b283-0030488c597c