Python API changelog r37840-40524 (HEAD) updating the doc.
I'm also including the latest API dump with it. This helps to keep track of the revisions to update and the next person updating doesnt need to build an old Blender for that (as I just had to).

I updated the sphinx_changelog_gen.py with this info as well.
I'll update changelog.rst manually next. I'm not so sure how to indicate that the properties were replaced instead of only removed.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@40525 954f8c5b-7b00-dc11-b283-0030488c597c